Large refactoring of functions from timeservice.go into utils.go
